+++ packages/orq-rc/RELEASES.md
@@ -10749,3 +10749,13 @@ Based on:
 - [typescript v4.12.0-rc.19] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.19] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.19 - packages/orq-rc
+
+## 2026-07-14 08:38:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.0 (2.917.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.20] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.20 - packages/orq-rc